documentation

+++ Controllers/ReviewerController.cs
@@ -10,10 +10,9 @@ namespace CSHARPAPI_WineReview.Controllers
 
     [ApiController]
     [Route("api/v1/[controller]")]
-    //dependency injection
+    // Dependency Injection
     public class ReviewerController(WineReviewContext context, IMapper mapper) : WineReviewController(context, mapper)
     {
-       
         /// <summary>
         /// Get all reviewers from the table.
         /// </summary>
@@ -36,39 +35,36 @@ namespace CSHARPAPI_WineReview.Controllers
         }
 
         /// <summary>
-        /// Get a reviewer by ID from the table.
+        /// Get a reviewer by ID.
         /// </summary>
         /// <param name="id">The ID of the reviewer.</param>
-        /// <returns>The reviewer with the specified ID.</returns>
+        /// <returns>A reviewer.</returns>
         [HttpGet("{id:int}")]
-        public ActionResult<List<ReviewerDTORead>> GetById(int id)
+        public ActionResult<ReviewerDTORead> GetById(int id)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(new { message = ModelState });
             }
-
-            Reviewer? r;
             try
             {
-                r = _context.Reviewers.FirstOrDefault(r => r.Id == id);
+                var reviewer = _context.Reviewers.Find(id);
+                if (reviewer == null)
+                {
+                    return NotFound();
+                }
+                return Ok(_mapper.Map<ReviewerDTORead>(reviewer));
             }
             catch (Exception ex)
             {
                 return BadRequest(new { message = ex.Message });
             }
-            if (r == null)
-            {
-                return NotFound(new { message = "Recenzent ne postoji u bazi" });
-            }
-
-            return Ok(_mapper.Map<ReviewerDTORead>(r));
         }
 
         /// <summary>
-        /// Create a new reviewer entry in the table.
+        /// Create a new reviewer.
         /// </summary>
-        /// <param name="reviewer">The reviewer to create.</param>
+        /// <param name="dto">The reviewer data transfer object.</param>
         /// <returns>The created reviewer.</returns>
         [HttpPost]
         public IActionResult Post(ReviewerDTOInsertUpdate dto)
@@ -77,17 +73,12 @@ namespace CSHARPAPI_WineReview.Controllers
             {
                 return BadRequest(new { message = ModelState });
             }
-
-            Reviewer? r;
-
-            //creates new entry in Reviewer table
             try
             {
-
-                r = _mapper.Map<Reviewer>(dto);
-                _context.Reviewers.Add(r);               
+                var reviewer = _mapper.Map<Reviewer>(dto);
+                _context.Reviewers.Add(reviewer);
                 _context.SaveChanges();
-                return StatusCode(StatusCodes.Status201Created, _mapper.Map<Reviewer>(r));
+                return CreatedAtAction(nameof(GetById), new { id = reviewer.Id }, _mapper.Map<ReviewerDTORead>(reviewer));
             }
             catch (Exception ex)
             {
@@ -96,11 +87,11 @@ namespace CSHARPAPI_WineReview.Controllers
         }
 
         /// <summary>
-        /// Update an existing reviewer entry in the table.
+        /// Update an existing reviewer.
         /// </summary>
         /// <param name="id">The ID of the reviewer to update.</param>
-        /// <param name="reviewer">The updated reviewer data.</param>
-        /// <returns>A status indicating the result of the update.</returns>
+        /// <param name="dto">The reviewer data transfer object.</param>
+        /// <returns>No content.</returns>
         [HttpPut("{id:int}")]
         [Produces("application/json")]
         public IActionResult Put(int id, ReviewerDTOInsertUpdate dto)
@@ -109,77 +100,80 @@ namespace CSHARPAPI_WineReview.Controllers
             {
                 return BadRequest(new { message = ModelState });
             }
-
-            Reviewer? r;
-            //updates existing entry retrieve by ID in Wine table
             try
             {
-                r = _context.Reviewers.FirstOrDefault(r => r.Id == id);
+                var reviewer = _context.Reviewers.Find(id);
+                if (reviewer == null)
+                {
+                    return NotFound();
+                }
+                _mapper.Map(dto, reviewer);
+                _context.SaveChanges();
+                return NoContent();
             }
             catch (Exception ex)
             {
                 return BadRequest(new { message = ex.Message });
             }
-            r = _mapper.Map(dto, r);
-            _context.Reviewers.Update(r);
-
-            _context.SaveChanges();
-            return StatusCode(StatusCodes.Status201Created, new { message = "Uspješno promijenjeno", wine = _mapper.Map<Reviewer>(r) });
         }
 
         /// <summary>
-        /// Delete a reviewer by ID from the table.
+        /// Delete a reviewer by ID.
         /// </summary>
         /// <param name="id">The ID of the reviewer to delete.</param>
-        /// <returns>A status indicating the result of the delete operation.</returns>
+        /// <returns>No content.</returns>
         [HttpDelete("{id:int}")]
         [Produces("application/json")]
         public async Task<IActionResult> Delete(int id)
         {
-            var reviewer = await _context.Reviewers.FindAsync(id);
-            if (reviewer == null)
+            if (!ModelState.IsValid)
             {
-                return NotFound(new { message = "Korisnik nije pronađen" });
+                return BadRequest(new { message = ModelState });
             }
-
-            //unable to delete data - handle 500 internal server error
             try
             {
+                var reviewer = await _context.Reviewers.FindAsync(id);
+                if (reviewer == null)
+                {
+                    return NotFound();
+                }
                 _context.Reviewers.Remove(reviewer);
                 await _context.SaveChangesAsync();
+                return NoContent();
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                return StatusCode(StatusCodes.Status500InternalServerError,
-                   new { message = "nije moguće obrisati, korisnik je napravio recenziju" });
+                return BadRequest(new { message = ex.Message });
             }
-
-            return Ok(new { message = "Uspješno obrisano" });
         }
 
+        /// <summary>
+        /// Get reviewers by pages.
+        /// </summary>
+        /// <param name="page">The page number.</param>
+        /// <param name="condition">The condition to filter reviewers.</param>
+        /// <returns>A list of reviewers.</returns>
         [HttpGet]
         [Route("getPages/{page}")]
         public IActionResult GetPages(int page, string condition = "")
         {
-            var byPage = 4;
-            condition= condition.ToLower();
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(new { message = ModelState });
+            }
             try
             {
                 var reviewers = _context.Reviewers
-                    .Where(r => EF.Functions.Like(r.FirstName.ToLower(), "%" + condition + "%")
-                    || EF.Functions.Like(r.LastName.ToLower(), "%" + condition + "%"))
-                    .Skip((byPage * page) - byPage)
-                    .Take(byPage)
+                    .Where(r => string.IsNullOrEmpty(condition) || r.FirstName.Contains(condition) || r.LastName.Contains(condition))
+                    .Skip((page - 1) * 10)
+                    .Take(10)
                     .ToList();
-
                 return Ok(_mapper.Map<List<ReviewerDTORead>>(reviewers));
             }
             catch (Exception ex)
             {
-
-                return BadRequest(ex.Message);  
+                return BadRequest(new { message = ex.Message });
             }
         }
-
     }
 }
